ReferralYear instead of MsurYear for longitudinal graph
I decided to address the ‘year’ issue before the ‘subtype’ issue.  This
graph uses the ReferralYear instead of the MsurYear.  It sounds like an
easy change, but the decision/programming went all the way back to the
PHI manipulation.

We knew MsurYear wasn’t ideal, but I didn’t expect this big of a
difference.  Here are some examples I see so far:
-The state trend is roughly the same, but it starts higher  and peaks
lower.
-The county peaks are much lower (around .28 instead of .48)
-Many of the same peaking counties appear in successive years now.
Before they typically would go from zero to <large> back to zero.